+++ src/main/java/graphql/GraphQL.java
@@ -602,18 +602,4 @@ public class GraphQL {
         }
         return new ChainedInstrumentation(instrumentationList);
     }
-
-
-    private static boolean USE_LIGHTWEIGHT_DFS = true;
-
-    @Internal
-    public static boolean lightWeightDataFetching() {
-        return USE_LIGHTWEIGHT_DFS;
-    }
-
-    @Internal
-    public static void useLightWeightDataFetching(boolean flag) {
-        USE_LIGHTWEIGHT_DFS = flag;
-    }
-
 }
